# src/main/apiServer/generated/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cherry Studio API",
    "version": "1.0.0",
    "description": "OpenAI-compatible API for Cherry Studio with additional Cherry-specific endpoints",
    "contact": {
      "name": "Cherry Studio",
      "url": "https://github.com/CherryHQ/cherry-studio"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use the API key from Cherry Studio settings"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "AssistantModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Model identifier",
            "example": "gpt-4"
          },
          "provider": {
            "type": "string",
            "description": "Provider identifier",
            "example": "openai"
          },
          "name": {
            "type": "string",
            "description": "Display name",
            "example": "GPT-4"
          }
        }
      },
...
      "AssistantDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/AssistantModel"
          },
          "settings": {
            "type": "object",
            "description": "Assistant settings (contextCount, temperature, etc.)"
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicSummary"
            }
          }
        }
      },
...
      "TopicDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "assistantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pinned": {
            "type": "boolean"
          }
        }
      },
      "CreateTopicRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Topic name (optional, defaults to \"New Topic\")",
            "example": "My Conversation"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/assistants/{assistantId}": {
      "get": {
        "summary": "Get assistant details",
        "description": "Returns a single assistant with full details including settings",
        "tags": ["Assistants"],
        "parameters": [
          {
            "in": "path",
            "name": "assistantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Assistant ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Assistant details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantDetail"
                }
              }
            }
          },
          "404": {
            "description": "Assistant not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistants/{assistantId}/topics": {
      "post": {
        "summary": "Create a new topic",
        "description": "Creates a new conversation topic under the specified assistant. The topic will appear in the UI immediately.",
        "tags": ["Assistants"],
        "parameters": [
          {
            "in": "path",
            "name": "assistantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Assistant ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTopicRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Topic created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDetail"
                }
              }
            }
          },
          "404": {
            "description": "Assistant not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List topics for an assistant",
        "description": "Returns all topics belonging to the specified assistant",
        "tags": ["Assistants"],
        "parameters": [
          {
            "in": "path",
            "name": "assistantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Assistant ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of topics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopicDetail"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Assistant not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistants/{assistantId}/topics/{topicId}": {
      "get": {
        "summary": "Get topic details",
        "description": "Returns details of a specific topic",
        "tags": ["Assistants"],
        "parameters": [
          {
            "in": "path",
            "name": "assistantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Assistant ID"
          },
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Topic ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Topic details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDetail"
                }
              }
            }
          },
          "404": {
            "description": "Assistant or topic not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistants/{assistantId}/topics/{topicId}/chat": {
      "post": {
        "summary": "Chat in a topic (SSE streaming)",
        "description": "Sends a user message and streams the AI assistant response via Server-Sent Events.\nThe message history is automatically built from the topic's existing messages.\nBoth user and assistant messages are persisted and will appear in the UI immediately.\nMCP tools configured on the assistant are automatically included and executed.\n\n**SSE event types:**\n```\ndata: {\"type\":\"text\",\"content\":\"Hello\",\"message_id\":\"uuid\"}\\n\\n\ndata: {\"type\":\"tool_call\",\"tool_call_id\":\"tc1\",\"name\":\"server__tool\",\"arguments\":{...}}\\n\\n\ndata: {\"type\":\"tool_result\",\"tool_call_id\":\"tc1\",\"content\":\"result...\"}\\n\\n\ndata: {\"type\":\"text\",\"content\":\" world\",\"message_id\":\"uuid\"}\\n\\n\ndata: [DONE]\\n\\n\n```\n",
        "tags": ["Assistants"],
        "parameters": [
          {
            "in": "path",
            "name": "assistantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Assistant ID"
          },
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Topic ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of assistant response events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStreamEvent"
                }
              }
            }
          },
          "400": {
            "description": "No model configured on assistant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Assistant or topic not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistants/{assistantId}/topics/{topicId}/messages": {
      "get": {
        "summary": "List messages in a topic",
        "description": "Returns all messages in a topic with their content blocks",
        "tags": ["Assistants"],
        "parameters": [
          {
            "in": "path",
            "name": "assistantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Assistant ID"
          },
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Topic ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages with blocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MessageDetail"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Assistant or topic not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}